+++ spec/lib/aixm/document_spec.rb
@@ -19,6 +19,21 @@ describe AIXM::Document do
     end
   end
 
+  describe :sourced_at= do
+    it "fails on invalid values" do
+      _(['foobar', '2018-01-77']).wont_be_written_to subject, :sourced_at
+    end
+
+    it "accepts nil" do
+      _([nil]).must_be_written_to subject, :sourced_at
+    end
+
+    it "parses dates and times" do
+      string = '2018-01-01 12:00:00 +0000'
+      _(subject.tap { _1.sourced_at = string }.sourced_at).must_equal Time.parse(string)
+    end
+  end
+
   describe :created_at= do
     it "fails on invalid values" do
       _(['foobar', '2018-01-77']).wont_be_written_to subject, :created_at
@@ -1128,7 +1143,7 @@ describe AIXM::Document do
     it "builds correct OFMX" do
       _(subject.to_xml).must_equal <<~"END"
         <?xml version="1.0" encoding="UTF-8"?>
-        <OFMX-Snapshot xmlns:xsi="http://schema.openflightmaps.org/0.2/OFMX-Snapshot.xsd" version="0" origin="rubygem aixm-#{AIXM::VERSION}" namespace="00000000-0000-0000-0000-000000000000" regions="EG LF" created="2022-04-21T00:00:00Z" effective="2022-04-21T00:00:00Z" expiration="2022-05-18T23:59:59Z">
+        <OFMX-Snapshot xmlns:xsi="http://schema.openflightmaps.org/0.2/OFMX-Snapshot.xsd" version="0" origin="rubygem aixm-#{AIXM::VERSION}" namespace="00000000-0000-0000-0000-000000000000" regions="EG LF" sourced="2022-04-20T00:00:00Z" created="2022-04-21T00:00:00Z" effective="2022-04-21T00:00:00Z" expiration="2022-05-18T23:59:59Z">
           <!-- Organisation: FRANCE -->
           <Org source="LF|GEN|0.0 FACTORY|0|0">
             <OrgUid region="LF">
@@ -2206,7 +2221,7 @@ describe AIXM::Document do
       AIXM.config.mid = true
       _(subject.to_xml).must_equal <<~"END"
         <?xml version="1.0" encoding="UTF-8"?>
-        <OFMX-Snapshot xmlns:xsi="http://schema.openflightmaps.org/0.2/OFMX-Snapshot.xsd" version="0" origin="rubygem aixm-#{AIXM::VERSION}" namespace="00000000-0000-0000-0000-000000000000" regions="EG LF" created="2022-04-21T00:00:00Z" effective="2022-04-21T00:00:00Z" expiration="2022-05-18T23:59:59Z">
+        <OFMX-Snapshot xmlns:xsi="http://schema.openflightmaps.org/0.2/OFMX-Snapshot.xsd" version="0" origin="rubygem aixm-#{AIXM::VERSION}" namespace="00000000-0000-0000-0000-000000000000" regions="EG LF" sourced="2022-04-20T00:00:00Z" created="2022-04-21T00:00:00Z" effective="2022-04-21T00:00:00Z" expiration="2022-05-18T23:59:59Z">
           <!-- Organisation: FRANCE -->
           <Org source="LF|GEN|0.0 FACTORY|0|0">
             <OrgUid region="LF" mid="971ba0a9-3714-12d5-d139-d26d5f1d6f25">